feat: build tracker remove build from checklist
Removed the dialog for confirming the removal of a build from the checklist and replaced it with a custom action dialog.
Implemented logic to update the checklist item count in the ListsInformationViewModel after removing a build.
Added a function to update the list of builds in the RecyclerView adapter and notify the adapter of changes.
Updated the layout of the checklist item card, renaming btn_viewChecklist to btn_addToChecklist.