feat: landing page username display improvement
The changes modify how the username is displayed on the landing page. Instead of displaying the full username, it will now truncate usernames longer than 13 characters, appending "..." to the truncated text. If the username is null, it will display "BuzyPerson". Also, the text size of "Hello user" was updated to use a dimension instead of hard-coding the size.