feat: added ic_photo_library and ic_photo_camera vector drawables
These files are defined as vector graphics in XML, designed to be scalable without quality loss. The `ic_photo_library` drawable depicts a library of photos, and the `ic_photo_camera` drawable represents a camera. Both utilize paths for drawing shapes and are configured with a tint attribute to allow for color adjustments based on the current theme.